# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sports-Point API",
    "version": "1.0.0",
    "description": "运动会积分与成绩管理系统对外 API 文档。"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/athletes": {
      "get": {
        "tags": [
          "运动员"
        ],
        "summary": "查询运动员",
        "description": "按类型和关键词查询运动员列表。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Athlete"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "page_size": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "query",
            "required": false,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "description": "运动员号、姓名或单位关键词",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "运动员"
        ],
        "summary": "新增运动员",
        "description": "按单位名称新增竞技或趣味运动员。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAthleteRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateAthleteRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/athletes/delete": {
      "post": {
        "tags": [
          "运动员"
        ],
        "summary": "删除运动员（表单兼容）",
        "description": "通过运动员类型和运动员号删除运动员。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAthleteRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAthleteRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/athletes/{athlete_type}/{athlete_no}": {
      "delete": {
        "tags": [
          "运动员"
        ],
        "summary": "删除运动员",
        "description": "通过路径参数删除运动员。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "path",
            "required": true,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_no",
            "in": "path",
            "required": true,
            "description": "运动员号",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/athletes/{athlete_type}/{athlete_no}/registrations": {
      "get": {
        "tags": [
          "运动员报名"
        ],
        "summary": "查询运动员已报名项目",
        "description": "查询指定运动员已经报名的个人项目。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registration"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "path",
            "required": true,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_no",
            "in": "path",
            "required": true,
            "description": "运动员号",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/athletes/registered-events": {
      "get": {
        "tags": [
          "运动员报名"
        ],
        "summary": "查询运动员已报名项目（兼容）",
        "description": "通过查询参数查询指定运动员已经报名的个人项目。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Registration"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "query",
            "required": true,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_no",
            "in": "query",
            "required": true,
            "description": "运动员号",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/athletes/{athlete_type}/{athlete_no}/registrations/{event_id}": {
      "post": {
        "tags": [
          "运动员报名"
        ],
        "summary": "增加运动员报名",
        "description": "为指定运动员增加一个个人项目报名。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "path",
            "required": true,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_no",
            "in": "path",
            "required": true,
            "description": "运动员号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "delete": {
        "tags": [
          "运动员报名"
        ],
        "summary": "删除运动员报名",
        "description": "移除指定运动员的一个个人项目报名。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "athlete_type",
            "in": "path",
            "required": true,
            "description": "运动员类型：competitive/fun",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "athlete_no",
            "in": "path",
            "required": true,
            "description": "运动员号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "description": "项目 ID",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/api/v1/athletes/registrations/add": {
      "post": {
        "tags": [
          "运动员报名"
        ],
        "summary": "增加运动员报名（表单兼容）",
        "description": "通过表单或 JSON 为运动员增加个人项目报名。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/athletes/registrations/remove": {
      "post": {
        "tags": [
          "运动员报名"
        ],
        "summary": "删除运动员报名（表单兼容）",
        "description": "通过表单或 JSON 移除运动员个人项目报名。",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "参数错误"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
      "Athlete": {
        "type": "object",
        "properties": {
          "age_group": {
            "type": "string",
            "description": "组别"
          },
          "athlete_no": {
            "type": "string",
            "description": "运动员号"
          },
          "athlete_ref_id": {
            "type": "integer",
            "description": "运动员数据库 ID"
          },
          "athlete_type": {
            "type": "string",
            "description": "运动员类型：competitive/fun"
          },
          "department_name": {
            "type": "string",
            "description": "归属单位"
          },
          "gender": {
            "type": "string",
            "description": "性别：male/female"
          },
          "name": {
            "type": "string",
            "description": "姓名"
          },
          "registered_events": {
            "type": "string",
            "description": "已报名项目列表"
          },
          "registration_count": {
            "type": "integer",
            "description": "已报名项目数"
          }
        }
      },
      "Registration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "项目 ID"
          },
          "name": {
            "type": "string",
            "description": "项目名称"
          },
          "label": {
            "type": "string",
            "description": "项目完整显示名"
          },
          "age_group": {
            "type": "string",
            "description": "组别"
          },
          "gender": {
            "type": "string",
            "description": "性别"
          }
        }
      },
...
      "CreateAthleteRequest": {
        "type": "object",
        "properties": {
          "athlete_type": {
            "type": "string",
            "description": "运动员类型：competitive/fun"
          },
          "athlete_no": {
            "type": "string",
            "description": "运动员号"
          },
          "name": {
            "type": "string",
            "description": "姓名"
          },
          "gender": {
            "type": "string",
            "description": "性别：male/female"
          },
          "department_name": {
            "type": "string",
            "description": "归属单位"
          },
          "age_group": {
            "type": "string",
            "description": "年龄组，可为空"
          }
        },
        "required": [
          "athlete_type",
          "athlete_no",
          "name",
          "gender",
          "department_name"
        ]
      },
      "DeleteAthleteRequest": {
        "type": "object",
        "properties": {
          "athlete_type": {
            "type": "string",
            "description": "运动员类型：competitive/fun"
          },
          "athlete_no": {
            "type": "string",
            "description": "运动员号"
          }
        },
        "required": [
          "athlete_type",
          "athlete_no"
        ]
      },
      "RegistrationRequest": {
        "type": "object",
        "properties": {
          "athlete_type": {
            "type": "string",
            "description": "运动员类型：competitive/fun"
          },
          "athlete_no": {
            "type": "string",
            "description": "运动员号"
          },
          "event_id": {
            "type": "integer",
            "description": "项目 ID"
          }
        },
        "required": [
          "athlete_type",
          "athlete_no",
          "event_id"
        ]
      },
...
    }
  }
}